fix: 5个Agent最终修复 - 编译/测试/警告/依赖
## 修复内容

### 1. easyssh-winui编译错误 ✅
- Shadow::small_dark() 修复
- linear_multiply类型修复
- Match arms类型修复
- Borrow错误修复
- Literal范围修复
- SftpEntry类型添加
- 结果: 零错误

### 2. 测试失败修复 ✅
- 从46个减少到34个
- 修复12个测试:
  - db::test_new_group_deserialization
  - edition::test_version_comparator
  - edition::test_supported_features
  - version::test_full_build_info_to_json
  - version_ffi::test_build_info_json
  - update_checker::test_is_asset_for_platform
  - config::manager (5个tokio测试)
  - debug::logging::test_log_filter

### 3. Clippy警告清理 ✅
- should_implement_trait警告
- arc_with_non_send_sync警告
- redundant_allocation警告
- too_many_arguments警告
- type_complexity警告
- dead_code警告
- 结果: 零警告

### 4. 依赖版本更新 ✅
- lru: 0.12 -> 0.16.3
- glib: 0.19 -> 0.20.12
- 消除RUSTSEC警告

### 5. easyssh-core编译 ✅
- --features lite: 通过
- --features standard: 通过
- --features pro: 通过

+++ crates/easyssh-platforms/windows/easyssh-winui/src/dialogs.rs
@@ -234,7 +234,7 @@ impl GroupColorPicker {
                             // Checkmark for selected
                             if is_selected {
                                 let check_color =
-                                    if color_value.r() + color_value.g() + color_value.b() > 384 {
+                                    if color_value.r() as u16 + color_value.g() as u16 + color_value.b() as u16 > 384 {
                                         egui::Color32::BLACK
                                     } else {
                                         egui::Color32::WHITE
